authorizing actions using policies

+++ app/Http/Controllers/Api/V2/TaskController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
 use App\Http\Resources\TaskResource;
 use App\Models\Task;
+use Illuminate\Support\Facades\Gate;
 
 class TaskController extends Controller
 {
@@ -16,8 +17,7 @@ class TaskController extends Controller
      */
     public function index()
     {
-        // return TaskResource::collection(Task::all());
-
+        Gate::authorize('viewAny', Task::class);
         return request()->user()->tasks()->get()->toResourceCollection();
     }
 
@@ -34,6 +34,7 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
+        Gate::authorize('create', Task::class);
         // $task = Task::create($request->validated());
 
         $task = $request->user()->tasks()->create($request->validated());
@@ -45,6 +46,7 @@ class TaskController extends Controller
      */
     public function show(Task $task)
     {
+        Gate::authorize('view', $task);
         // return new TaskResource($task);
         // return TaskResource::make($task);
         return $task->toResource();
@@ -55,6 +57,8 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        Gate::authorize('update', $task);
+
         $task->update($request->validated());
         return $task->toResource();
     }
@@ -64,8 +68,9 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
-        $task->delete();
+        Gate::authorize('delete', $task);
 
+        $task->delete();
         return response()->noContent();
     }
 }
